+++ src/main/java/com/starbucks/back/auth/application/AuthService.java
@@ -4,6 +4,8 @@ import com.starbucks.back.auth.dto.in.RequestSignInDto;
 import com.starbucks.back.auth.dto.in.RequestSignUpDto;
 import com.starbucks.back.auth.dto.out.ResponseSignInDto;
 import com.starbucks.back.user.domain.User;
+import com.starbucks.back.auth.dto.in.RequestGetUserNicknameDto;
+import com.starbucks.back.auth.dto.out.ResponseGetUserNicknameDto;
 
 public interface AuthService {
     void signUp(RequestSignUpDto requestSignUpDto);
@@ -14,4 +16,5 @@ public interface AuthService {
     boolean existsNickname(String nickname);
     boolean existsPhoneNumber(String phoneNumber);
     void oauthSignUp(User user);
+    ResponseGetUserNicknameDto getUserNickname(RequestGetUserNicknameDto requestGetUserNicknameDto);
 }
