+++ src/main/java/com/starbucks/back/user/application/UserServiceImpl.java
@@ -5,14 +5,11 @@ import com.starbucks.back.common.exception.BaseException;
 import com.starbucks.back.common.util.RedisUtil;
 import com.starbucks.back.oauth.application.OauthService;
 import com.starbucks.back.user.domain.User;
-import com.starbucks.back.user.domain.enums.UserState;
 import com.starbucks.back.user.dto.in.*;
 import com.starbucks.back.user.dto.out.ResponseGetUserInfoDto;
-import com.starbucks.back.user.dto.out.ResponseGetUserNicknameDto;
 import com.starbucks.back.user.infrastructure.UserRepository;
 import com.starbucks.back.user_withdrwal_pending.application.UserWithdrawalPendingService;
 import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -54,15 +51,6 @@ public class UserServiceImpl implements UserService{
                 userRepository.findByUserUuid(userUuid).orElseThrow(
                         () -> new IllegalArgumentException(userUuid)
                 )
-        );
-    }
-
-    @Override
-    public ResponseGetUserNicknameDto getUserNickname(RequestGetUserNicknameDto requestGetUserNicknameDto) {
-        return ResponseGetUserNicknameDto.from(
-                userRepository.findByUserUuid(requestGetUserNicknameDto.getUserUuid())
-                        .orElseThrow(() -> new BaseException(BaseResponseStatus.NO_EXIST_USER))
-                        .getNickname()
         );
     }
