+++ src/main/java/com/starbucks/back/user/presentation/UserController.java
@@ -1,14 +1,14 @@
 package com.starbucks.back.user.presentation;
 
+import com.starbucks.back.auth.dto.in.RequestGetUserNicknameDto;
 import com.starbucks.back.common.entity.BaseResponseEntity;
 import com.starbucks.back.common.entity.BaseResponseStatus;
 import com.starbucks.back.common.util.SecurityUtil;
 import com.starbucks.back.user.application.UserService;
 import com.starbucks.back.user.dto.in.*;
-import com.starbucks.back.user.dto.out.ResponseGetUserNicknameDto;
 import com.starbucks.back.user.vo.in.*;
 import com.starbucks.back.user.vo.out.ResponseGetUserInfoVo;
-import com.starbucks.back.user.vo.out.ResponseGetUserNicknameVo;
+import com.starbucks.back.auth.vo.out.ResponseGetUserNicknameVo;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -36,16 +36,6 @@ public class UserController {
         System.out.println(securityUtil.getCurrentUserNickname());
         return new BaseResponseEntity<>(
                 securityUtil.getCurrentUserDetails()
-        );
-    }
-
-    @Operation(summary = "Get User Nickname API", description = "유저 닉네임 조회", tags = {"User-service"})
-    @GetMapping("/nickname")
-    public BaseResponseEntity<ResponseGetUserNicknameVo> getUserNickname(
-            @RequestHeader(value = "Uuid") String userUuid
-    ) {
-        return new BaseResponseEntity<>(
-                userService.getUserNickname(RequestGetUserNicknameDto.from(userUuid)).toVo()
         );
     }
